# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      err_code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
...
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.SecurityEvidenceMetadata:
    properties:
      clusterType:
        example: ablestack-hci
        type: string
      collectedHosts:
        items:
          type: string
        type: array
      collectorStatus:
        type: integer
      collectorWarning:
        type: string
      downloadAvailable:
        type: boolean
      filename:
        example: ABLESTACK 보안 취약점 증적 자료.zip
        type: string
      generatedAt:
        example: "2026-07-27T10:30:00+09:00"
        type: string
      hosts:
        example: 3
        type: integer
      items:
        example: 67
        type: integer
      path:
        type: string
      requestedHosts:
        example: 3
        type: integer
      requestedTargets:
        items:
          type: string
        type: array
      runDirectory:
        type: string
      size:
        example: 1048576
        type: integer
      slides:
        example: 68
        type: integer
      targetGroups:
        items:
          type: string
        type: array
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.SecurityEvidenceRequest:
    properties:
      hosts:
        description: cluster.json 대신 명시적으로 수집할 hostname/IP 목록
        example:
        - ablecube1
        - scvm1
        items:
          type: string
        type: array
      items:
        description: '점검 항목: all, U-01,U-02 또는 U-01:U-13'
        example: all
        type: string
      max_output_lines:
        description: 명령별 최대 출력 행 수
        example: 400
        type: integer
      ssh_port:
        description: SSH 포트. 0이면 로컬 sshd 설정에서 자동 감지한다.
        example: 22
        type: integer
      ssh_user:
        description: SSH 사용자
        example: root
        type: string
      targets:
        description: '대상 그룹: all/ablecube/scvm/ccvm'
        example:
        - all
        items:
          type: string
        type: array
      timeout:
        description: 호스트별 명령 제한 시간(초)
        example: 120
        type: integer
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.SecurityEvidenceResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: ok
        type: string
      val:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.SecurityEvidenceMetadata'
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
...
  /cube/security/evidence:
    get:
      description: 가장 최근에 생성된 보안 증적 ZIP 메타데이터를 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.SecurityEvidenceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Latest Security Evidence
      tags:
      - Cube-Security
    post:
      consumes:
      - application/json
      description: cluster.json 대상의 읽기 전용 보안 점검 결과를 수집해 TXT/XLSX/PPTX가 포함된 ZIP 패키지를
        생성합니다.
      parameters:
      - description: security evidence options
        in: body
        name: body
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.SecurityEvidenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.SecurityEvidenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Generate Security Evidence
      tags:
      - Cube-Security
  /cube/security/evidence/download:
    get:
      description: 가장 최근에 생성된 보안 증적 ZIP 파일을 다운로드합니다.
      produces:
      - application/zip
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Download Security Evidence
      tags:
      - Cube-Security
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"